Fix angle data type for circular grid grounds
The angle must be entered as a float.  In the inx file this is correctly specified but in the python file the angle is specified as an integer.  In past versions of Inkscape, the inx file must have overridden the python description but that is no longer the case with Inkscape 1.4.